+++ .github/workflows/ci.yml
@@ -30,11 +30,11 @@ jobs:
       - name: rustfmt
         run: |
           cd zygisk && cargo fmt --check
-          cd ../test-app/native && cargo fmt --check
+          cd ../lsposed/native && cargo fmt --check
       - name: clippy (zygisk)
         run: cd zygisk && cargo ndk -t arm64-v8a clippy -- -D warnings
-      - name: clippy (test-app)
-        run: cd test-app/native && cargo ndk -t arm64-v8a clippy -- -D warnings
+      - name: clippy (lsposed native)
+        run: cd lsposed/native && cargo ndk -t arm64-v8a clippy -- -D warnings
       - name: cargo test (zygisk)
         run: cd zygisk && cargo test
 
@@ -44,7 +44,7 @@ jobs:
 
       # Kotlin
       - name: ktlint
-        run: ktlint "lsposed/**/*.kt" "test-app/**/*.kt"
+        run: ktlint "lsposed/**/*.kt"
 
   kmod:
     runs-on: ubuntu-latest
@@ -168,63 +168,8 @@ jobs:
           path: vpnhide-lsposed.apk
           if-no-files-found: error
 
-  test-app:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set up JDK 17
-        uses: actions/setup-java@v4
-        with:
-          distribution: temurin
-          java-version: '17'
-
-      - name: Set up Gradle
-        uses: gradle/actions/setup-gradle@v4
-
-      - name: Install Rust + cargo-ndk
-        run: |
-          curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --default-toolchain stable --profile minimal
-          . "$HOME/.cargo/env"
-          rustup target add aarch64-linux-android
-          cargo install cargo-ndk --locked
-
-      - name: Set up Android NDK
-        uses: nttld/setup-ndk@v1
-        id: ndk
-        with:
-          ndk-version: r28b
-
-      - name: Set up keystore
-        env:
-          KEYSTORE_BASE64: ${{ secrets.ANDROID_KEYSTORE_BASE64 }}
-          KEYSTORE_PASSWORD: ${{ secrets.ANDROID_KEYSTORE_PASSWORD }}
-          KEY_ALIAS: ${{ secrets.ANDROID_KEY_ALIAS }}
-        run: |
-          KEYSTORE_PATH="${{ github.workspace }}/test-app/release.jks"
-          echo "$KEYSTORE_BASE64" | base64 --decode > "$KEYSTORE_PATH"
-          cat > test-app/keystore.properties <<EOF
-          password=$KEYSTORE_PASSWORD
-          keyAlias=$KEY_ALIAS
-          storeFile=$KEYSTORE_PATH
-          EOF
-
-      - name: Build APK
-        run: |
-          . "$HOME/.cargo/env"
-          export ANDROID_NDK_HOME=${{ steps.ndk.outputs.ndk-path }}
-          cd test-app && ./gradlew --no-daemon assembleRelease
-          cp app/build/outputs/apk/release/app-release.apk "$GITHUB_WORKSPACE/vpnhide-test.apk"
-
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: vpnhide-test
-          path: vpnhide-test.apk
-          if-no-files-found: error
-
   release:
-    needs: [kmod, zygisk, lsposed, test-app]
+    needs: [kmod, zygisk, lsposed]
     if: startsWith(github.ref, 'refs/tags/v')
     runs-on: ubuntu-latest
     steps:
