--- test-app/app/src/main/java/dev/okhsunrog/vpnhide/test/NativeChecks.kt
@@ -1,39 +1,0 @@
-package dev.okhsunrog.vpnhide.test
-
-object NativeChecks {
-    init {
-        System.loadLibrary("vpnhide_test")
-    }
-
-    external fun checkIoctlSiocgifflags(): String
-
-    external fun checkIoctlSiocgifmtu(): String
-
-    external fun checkIoctlSiocgifconf(): String
-
-    external fun checkGetifaddrs(): String
-
-    external fun checkProcNetRoute(): String
-
-    external fun checkProcNetIfInet6(): String
-
-    external fun checkNetlinkGetlink(): String
-
-    external fun checkNetlinkGetroute(): String
-
-    external fun checkProcNetIpv6Route(): String
-
-    external fun checkProcNetTcp(): String
-
-    external fun checkProcNetTcp6(): String
-
-    external fun checkProcNetUdp(): String
-
-    external fun checkProcNetUdp6(): String
-
-    external fun checkProcNetDev(): String
-
-    external fun checkProcNetFibTrie(): String
-
-    external fun checkSysClassNet(): String
-}
