--- test-app/app/src/main/res/values-ru/strings.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name">VPN Hide Тест</string>
-    <string name="toolbar_title">VPN Hide Диагностика</string>
-
-    <!-- Summary -->
-    <string name="summary_running">Выполняется&#8230;</string>
-    <string name="summary_format">%1$d/%2$d пройдено</string>
-    <string name="btn_run_all">Запустить все проверки</string>
-
-    <!-- Badge labels -->
-    <string name="badge_pass">ОК</string>
-    <string name="badge_fail">ПРОВАЛ</string>
-    <string name="badge_info">ИНФО</string>
-
-    <!-- Banner -->
-    <string name="banner_warning">Включите VPN-соединение перед запуском проверок и убедитесь, что это приложение добавлено в список целевых в VPN Hide.</string>
-
-    <!-- Section headers -->
-    <string name="section_native">Нативный уровень (kmod / zygisk)</string>
-    <string name="section_java">Java API уровень (LSPosed)</string>
-</resources>
