--- test-app/native/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "vpnhide_test"
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.85"
-license = "MIT"
-
-[lib]
-name = "vpnhide_test"
-crate-type = ["cdylib"]
-
-[dependencies]
-jni = { version = "0.21", default-features = false }
-libc = "0.2"
-log = "0.4"
-android_logger = { version = "0.15", default-features = false }
-
-[profile.release]
-opt-level = "z"
-lto = "fat"
-codegen-units = 1
-strip = true
-panic = "abort"
-
-[profile.dev]
-panic = "abort"
